Programming/Python Group Project/requirements.txt
Flask
pyopenssl
request
bcrypt
regex
password-strength
sympy
pyjwt
flask_mail
itsdangerous
cryptography